+++ libgps.c
@@ -139,7 +139,7 @@ int gps_close(struct gps_data_t *gpsdata)
 }
 
 void gps_set_raw_hook(struct gps_data_t *gpsdata, 
-		      void (*hook)(struct gps_data_t *, char *, int len, int level))
+		      void (*hook)(struct gps_data_t *, char *, size_t len, int level))
 {
     gpsdata->raw_hook = hook;
 }
@@ -435,9 +435,9 @@ static void gps_unpack(char *buf, struct gps_data_t *gpsdata)
 
 /*@ -nullstate -compdef @*/
     if (gpsdata->raw_hook)
-	gpsdata->raw_hook(gpsdata, buf, (int)strlen(buf),  1);
+	gpsdata->raw_hook(gpsdata, buf, strlen(buf),  1);
     if (gpsdata->thread_hook)
-	gpsdata->thread_hook(gpsdata, buf, (int)strlen(buf), 1);
+	gpsdata->thread_hook(gpsdata, buf, strlen(buf), 1);
 }
 /*@ +nullstate +compdef @*/
 /*@ -branchstate +usereleased @*/
@@ -502,7 +502,7 @@ static void *poll_gpsd(void *args)
 }
 
 int gps_set_callback(struct gps_data_t *gpsdata, 
-		     void (*callback)(struct gps_data_t *sentence, char *buf, int len, int level),
+		     void (*callback)(struct gps_data_t *sentence, char *buf, size_t len, int level),
 		     pthread_t *handler) 
 /* set an asynchronous callback and launch a thread for it */
 {
